fix: 이용서약 라우트와 바뀐 atom 세팅
