+++ src/pages/common/applyPledgeStep/Step.tsx
@@ -1,7 +1,5 @@
 import { useAtomValue } from 'jotai';
 import FirstPage from './FirstPage';
-import { pageFinishAtom } from '~/store/funnel';
-import { useNavigate } from 'react-router-dom';
 import { meetingTypeAtom } from '~/store/meeting/common';
 import PageLayout from '~/components/layout/page/PageLayout';
 import { useFunnel } from '~/hooks/useFunnel';
@@ -9,8 +7,9 @@ import { useFunnel } from '~/hooks/useFunnel';
 const CommonBranchGatewayStep = () => {
   const { Funnel, currentPage, PageHandler } = useFunnel({
     pageNumberList: [1],
+    // deprecated
+    prevStep: { path: '/common/checkApplyInfoStep' },
     nextStep: { path: '/common/branchGatewayStep' },
-    prevStep: { path: '/' },
   });
   const meetingType = useAtomValue(meetingTypeAtom);
